Rename pkg to pi_spin
