+++ src/pi_spin/config.py
